requirements.txt
selenium==4.0.0
python-telegram-bot==13.7
webdriver-manager==3.8.6
